+++ contracts/btc-finality/src/contract.rs
@@ -1,10 +1,11 @@
 use crate::error::ContractError;
 use crate::finality::{
-    compute_active_finality_providers, handle_finality_signature, handle_public_randomness_commit,
-    handle_rewards_distribution, handle_unjail,
+    handle_finality_signature, handle_public_randomness_commit, handle_rewards_distribution,
+    handle_unjail,
 };
 use crate::liveness::handle_liveness;
 use crate::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use crate::power_dist_change::compute_active_finality_providers;
 use crate::state::config::{
     Config, ADMIN, CONFIG, DEFAULT_JAIL_DURATION, DEFAULT_MAX_ACTIVE_FINALITY_PROVIDERS,
     DEFAULT_MIN_PUB_RAND, DEFAULT_MISSED_BLOCKS_WINDOW, DEFAULT_REWARD_INTERVAL,
@@ -228,9 +229,9 @@ fn handle_update_staking(
 fn handle_begin_block(deps: &mut DepsMut, env: Env) -> Result<Response, ContractError> {
     // Compute active finality provider set
     let max_active_fps = CONFIG.load(deps.storage)?.max_active_finality_providers as usize;
-    compute_active_finality_providers(deps, &env, max_active_fps)?;
+    let response = compute_active_finality_providers(deps, &env, max_active_fps)?;
 
-    Ok(Response::new())
+    Ok(response)
 }
 
 fn handle_end_block(
@@ -263,7 +264,8 @@ fn handle_end_block(
 
     // Examine liveness of finality providers and jail them if they are inactive for a certain
     // duration.
-    handle_liveness(deps, &env, &cfg)?;
+    let liveness_events = handle_liveness(deps, &env, &cfg)?;
+    res = res.add_events(liveness_events);
 
     // On an reward interval boundary, send rewards for distribution to Babylon Genesis
     if env.block.height > 0 && env.block.height % cfg.reward_interval == 0 {